+++ vujes-estimation/app/Http/Controllers/ProjectController.php
@@ -4,23 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 
 class ProjectController extends Controller
 {
-    public function store(Request $request)
+    public function store(StoreProjectRequest $request)
     {
-        $validated = $request->validate([
-            'name' => [
-                'required',
-                'string',
-                Rule::unique('projects')->where('client_id', $request->client_id)],
-            'description' => 'nullable|string',
-            'client_id' => 'required|exists:clients,id']);
+        $validated = $request->validated();
 
-        Project::create($validated);
+        $project = Project::create($validated);
 
-        return response()->json(['message' => 'Project saved']);
+        return response()->json(['message' => 'Project saved', 'project'=>$project], 201);
 
     }
 
@@ -33,34 +28,27 @@ class ProjectController extends Controller
             ->with('client')
             ->get();
 
-        return response()->json($projects);
+        return response()->json($projects, 201);
 
     }
 
-    public function update(Request $request, $id)
+    public function update(UpdateProjectRequest $request, $id)
     {
         $project = Project::findOrFail($id);
-        $validated = $request->validate([
-            'name' => [
-                'required',
-                'string',
-                Rule::unique('projects')
-                    ->where('client_id', $request->client_id)
-                    ->ignore($project->id)],
-            'description' => 'nullable|string',
-            'client_id' => 'required|exists:clients,id',
-        ]);
+        Gate::authorize('update', $project);
+        $validated = $request->validated();
 
         $project->update($validated);
 
-        return response()->json(['message' => 'Project Updated']);
+        return response()->json(['message' => 'Project Updated', 'project'=>$project], 201);
     }
 
     public function destroy($id)
     {
         $project = Project::findOrFail($id);
+        Gate::authorize('delete', $project);
         $project->delete();
 
-        return response()->json(['message' => 'Project Deleted']);
+        return response()->json(['message' => 'Project Deleted'], 201);
     }
 }
